+++ packages/ui/src/components/widgets/OnrampWidget/modals/OnrampModal.tsx
@@ -423,8 +423,8 @@ export const OnrampModal: FC<OnrampModalProps> = ({
           isPassthrough
             ? toTokenTotalAmount
             : fromToken.symbol === 'ETH'
-              ? ethTotalAmount
-              : totalAmount ?? undefined
+            ? ethTotalAmount
+            : totalAmount ?? undefined
         }
         fiatCurrency={fiatCurrency}
         isPassthrough={isPassthrough}